Only compress if sum(weights) < floor(n/k)
This was inadvertently broken when we implemented exponential decay and make weightedCount a double instead of a long

+++ stats/src/main/java/com/proofpoint/stats/QuantileDigest.java
@@ -484,8 +484,7 @@ public class QuantileDigest
                     rightWeight = node.right.weightedCount;
                 }
 
-                boolean shouldCompress = node.weightedCount + leftWeight + rightWeight <
-                        weightedCount / compressionFactor;
+                boolean shouldCompress = node.weightedCount + leftWeight + rightWeight < (int) (weightedCount / compressionFactor);
 
                 double oldNodeWeight = node.weightedCount;
                 if (shouldCompress || leftWeight < ZERO_WEIGHT_THRESHOLD) {
